Spherical Projectiles
RAA now spawns a spherical projectile which works better.
Fixed bug in sphereCollider ray test.
Added collision layer mask system

+++ TBGrid/src/Scene.cpp
@@ -77,7 +77,7 @@ void Scene::addObject(std::shared_ptr<GameObject> object)
 void Scene::deleteObject(std::shared_ptr<GameObject> toDelete)
 {
 	//first recursively destroy the child objects
-	for (int i = toDelete->children.size() - 1; i >= 0; i--)
+	for (int i = static_cast<int>(toDelete->children.size()) - 1; i >= 0; i--)
 	{
 		deleteObject(toDelete->children[i]);
 	}
